web/requirements.txt
Flask
Flask-Admin==1.6.0
SQLAlchemy==1.3.16
flask-login
Flask-WTF
flask-bcrypt
cloudscraper
BeautifulSoup4
requests
Flask[async]
pytube
Flask-APScheduler
loguru
gunicorn
ffmpeg
click
fuzzywuzzy
python-Levenshtein
Flask-Caching
flask-migrate
python-dotenv
emoji
matplotlib
psutil
psycopg2